refactor: split file manager content routes
